+++ boards/matrix_portal_m4/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # v0.1.0
 
+- Added RedLedPwm alias to onboard led
 - Added blinky_basic.rs example
 - Added pin mapping to pin.rs